reworked stores and store owners, needs futher testing

+++ This_App_Backend/src/main/java/com/example/This_App_Backend/entity/Products.java
@@ -19,7 +19,7 @@ public class Products {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "product_id")
-    private Integer productId;
+    private Long productId;
 
     @ManyToOne // Many products can belong to one store
     @JoinColumn(name = "store_id", nullable = false) // Foreign key to Stores table
@@ -56,6 +56,7 @@ public class Products {
     private LocalDateTime updatedAt;
 
     // --- Relationships ---
+
 
     // One Product can be in many Order_Items
     @OneToMany(mappedBy = "product", cascade = CascadeType.ALL)
